day_3_api/requirements.txt
alembic==1.12.0
aniso8601==9.0.1
attrs==23.1.0
click==8.1.7
colorama==0.4.6
Flask==2.0.2
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==3.0.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
Mako==1.2.4
MarkupSafe==2.1.3
pkgutil-resolve-name==1.3.10
PyMySQL==1.0.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2023.3.post1
referencing==0.30.2
rpds-py==0.10.6
six==1.16.0
SQLAlchemy==1.4.39
typing-extensions==4.8.0
Werkzeug==2.0.2
zipp==3.17.0
